requirements.txt
#  ===========================================================================================================
#  File Name: requirements.txt
#  Description: This file lists out project's dependencies along with their versions.

#  -----------------------------------------------------------------------------------------------------------
#  Item Name: Whizhack Client Dashboard
#  Author URL: https://whizhack.in

#  ===========================================================================================================

amqp==5.1.1
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
boto3==1.26.42
botocore==1.29.42
celery==5.2.3
celery-progress==0.1.2
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.4
cycler==0.11.0
daphne==3.0.2
Deprecated==1.2.13
Django==4.0.5
django-celery-results==2.3.0
django-cors-headers==3.12.0
django-redis==5.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
fonttools==4.33.3
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==1.0.1
kiwisolver==1.4.3
kombu==5.2.4
matplotlib==3.5.2
numpy==1.23.0
opensearch-py==2.0.0
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
PyJWT==2.4.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.1.4
requests==2.28.1
s3transfer==0.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==22.4.0
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.10
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
zope.interface==5.4.0
